refactor: move main plugin class to core feature folder
