+++ packages/obsidian-plugin/src/features/core/types.ts
@@ -1,0 +1,8 @@
+import { type } from "arktype";
+import { sourceDocumentSettingsSchema } from "../source-document";
+
+const mcpToolsPluginSettingsSchema = type({
+  sourceDocument: sourceDocumentSettingsSchema,
+});
+
+export type McpToolsPluginSettings = typeof mcpToolsPluginSettingsSchema.infer;
