--- packages/obsidian-plugin/src/types.ts
@@ -1,12 +1,0 @@
-declare module "obsidian" {
-  interface McpToolsPluginSettings {
-    version?: string;
-  }
-
-  interface Plugin {
-    loadData(): Promise<McpToolsPluginSettings>;
-    saveData(data: McpToolsPluginSettings): Promise<void>;
-  }
-}
-
-export {};
